feat: reorg and rename + improve shortcode sample
