+++ Samples~/QuickStart/AvatarRender/Avatar Image Config.asset.meta
