+++ Samples~/QuickStart/AvatarRender/AvatarRenderSample.cs.meta
