+++ Samples~/QuickStart/AvatarRender/AvatarRenderSample.unity.meta
