+++ Samples~/QuickStart/AvatarShortCode.meta
