+++ Samples~/QuickStart/AvatarShortCode/Avatar Short Code Panel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1f44f5da77524b042a011f2eaf878fde
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
