+++ Samples~/QuickStart/AvatarShortCode/AvatarShortCode.unity.meta
