+++ Samples~/QuickStart/AvatarShortCode/AvatarShortCodeSample.cs.meta
